Define build info strings in a single module
