+++ nano/test_common/telemetry.cpp
@@ -1,3 +1,4 @@
+#include <nano/lib/version.hpp>
 #include <nano/node/endpoint.hpp>
 #include <nano/node/messages.hpp>
 #include <nano/node/node.hpp>
